+ option --no-bind to check on any address
Some application does not start on the local interfaces,
and/or it may start on some very low port. This option
is useful (when being used with --address option) to
check on any kind of `address:port`: our program `rolo`
will not bind on any interface; it just uses a simple
TCP connection to check if the port is open; if yes,
the `rolo` will report that your application is running.

Application: run an arbitary command if another problem
is running. For example, when using VBoxManage to start
some guests in "headless" mode, the guests may be open
some ports (rdesktop, ssh,...). The program VBoxManage
does not open any port by itself. To make sure that
we all have the guests to be running we use:

  rolo -a 1.2.3.4 -p 3389 --no-bind \
    VBoxManage startvm foobar --type headless

This is a real problem of my friend"s company ;)

In this commit there are also minor changes

* Update help messages
* Add new method (port_open?)
* Add "--timeout" option to use with "--no-bind"