Fix a crazy typo error in :close_on_exec
Because we do not print some error if Exception happens,
the typo error (:att instead of :attr) is simply ignored.
This happends for months and I do know really why things
are just working fine! This is may because Ruby-2.0 by
default set the flags close-on-exec => true, while the
other Rubies use :false as default. So crazy!